[Version] Update to 1.3.4 (#60)
* update ByteBuf indexOf
* set EventLoop memory limit to 2GB-63
* move channel.read channel.write to public
* add setBytes(long address)
* move Channel.read to Protocol
* limit idle time
* support HEAD method(http1.1)
* bug fix: check if unsafe available
* independent boot module

+++ firenio-core/src/main/java/com/firenio/buffer/ByteBufAllocator.java
@@ -28,7 +28,7 @@ public abstract class ByteBufAllocator extends LifeCycle {
 
     public abstract void freeMemory();
 
-    public abstract long getCapacity();
+    public abstract int getCapacity();
 
     public abstract int getUnit();
